refactor(IRGenerator): 重构函数元数据和局部变量处理
重构函数元数据(FuncMeta)结构，增加局部变量作用域信息
优化IRGenerator中局部变量的存储和查找方式
将作用域链改为基于PC的作用域管理
更新相关测试用例验证功能

+++ src/vm/Instruction.hpp
@@ -59,7 +59,7 @@ namespace Cial::Bytecode {
 
 namespace Cial::Bytecode::Op {
 
-    enum class OpCode : std::uint32_t {
+    enum class OpCode : std::uint16_t {
 #define OPCODE_ENUM_CLASS(OP) OP,
         OPCODE_ENUMS(OPCODE_ENUM_CLASS)
 #undef OPCODE_ENUM_CLASS
